Restructure tests
- Merge tests to `tests` directory.
- Change zuul to use those changes.
- Update `Tests` part in README.
