+++ zuul-tests.yaml
@@ -1,13 +1,11 @@
 ---
-- name: Build test image & run tests
+- name: Build test image & run tests in container
   hosts: all
   tasks:
-    - name: Install dependencies for tests
+    - name: Install git and podman
       dnf:
         name:
           - git-core
-          - packit
-          - python3-pytest
           - podman
       become: true
     - name: Set project_dir variable when running this locally
@@ -22,12 +20,10 @@
       args:
         chdir: "{{ project_dir }}"
       become: true
-    - name: pip-install dist2src for unit tests
-      pip:
-        name: "{{ project_dir }}"
-      become: true
     - name: Run tests
-      command: make check
+      command: make check-in-container
       args:
         chdir: "{{ project_dir }}"
+      environment:
+        COLOR: "no"
       become: true
